Fix hang when incoming p2p or group calls throw up dialog.
There were changes in atlasaurus that resulted in a hang for incoming p2p
and group calls which throw up dialogs.  The changes revolved around mutex,
coroutines, job queues, and such.
The fix was to do any processing that may result in callbacks from
the webrtc code in a queued job instead of a coroutine.