Allow UI to have lazily-loaded submodules.
Equip UI with an __index metamethod. When someone references an unknown
key/field in UI, require() that module and cache it for future reference.

Add util.setmetamethods() as a way to find or create a metatable on a
specified table containing specified metamethods.

Exercise the new functionality by referencing UI.popup in test_popup.lua.