Add context info to Develop->Render Tests->Frame Profile stats file.
Specifically, add the viewer version, the machine ID, the grid, the region
name and ID, the parcel name and ID and the timestamp. This is both richer and
less fragile than trying to extract that information from the generated
filename: e.g. we now have region and parcel names.

Instead of making `LLGLSLShader::finishProfile()` mess with file I/O, pass it
a reference to a `boost::json::value` to be filled in with statistics, if it's
a `boost::json::object`. Otherwise it's `boost::json::null`, meaning no report.

Make llviewerdisplay.cpp's `display()` function instantiate a `boost::json::value`
to pass to `finishProfile()`. That lets llviewerdisplay.cpp also set the
`"context"` entry, with a new `getProfileStatsContext()` function quite
similar to `getProfileStatsFilename()`.