set CURLOPT_ENCODING to nullptr in case of curl7
Curl 7 introduced more strictness how it handles CURLOPT_ENCODING.

With earlier versions it as apparently okay when the server sent an encoding curl did not implement.

Newer versions are much more stricter and refuse to complete downloads under such circumstances. This leads to almost nothing loading.
The culrpit seemingly being either the simulator or asset servers sending headers curls does not accept.

The workaround is to set CURLOPT_ENCODING to nullptr rather than "". This does come with the implication that curl will not handle decompression transparently, rather this burden gets shifted to the application.
It will also not send Accept-Encoding headers automatically.