Display the computed viewer_channel on build summary page.
Too often we've run a GH build to completion, only to discover that the
channel logic baked in the wrong channel and we must rerun. At least surface
the channel early in the build, when we still have a chance to cancel it and
rerun quickly.

Also engage the sign-pkg-windows update that eliminates warning spam.