Add liz-demo.lua script.
This demo script teleports to the Boston Lab platform, looks for a suitable
chair and walks over to it. It makes the avatar do a little dance, then (tries
to sit down on the chair) (sits near it). It then engages (a version of) the
Eliza program, conducted over local chat, until the user says "bye".