Add LLListener:inline() that suspends caller until done.
LLListener:start() implicitly launches a Lua coroutine that calls the passed
handler for every incoming event on the specified pump name, until either the
handler says it's done or LLListener:stop() is called.

But this is awkward for a calling script that wants to wait until the last of
the events on the specified pump name. Synchronizing requires something like
an additional WaitQueue.

Now a caller can invoke LLListener:inline(), which continues processing events
on the specified pump name until the pump is closed, or the handler says it's
done, or an error is raised. At that point control returns to the caller.

Also recast LLListener:start() in terms of leap.WaitForPump and
leap.consumeWaitFor(), the same underlying engine used by LLListener:inline().