Add LLAgent.localToGlobal(), globalToLocal() conversions.
Also add localToGlobalVector() and globalToLocalVector() that return Luau
vectors, instead of our conventional {x, y, z} tables. The idea is that you
can select the appropriate return type, instead of possibly requiring
unnecessary double conversions between table and vector.

These are in LLAgent because the Lua API we use to discover the world position
of the current region (the essential ingredient for these conversions) is the
'LLAgent' op 'getPosition', which returns both global and region positions.
The difference between these two is the corner of the current region.

Arguably we should instead use LLFloaterAbout.getInfo().POSITION and
POSITION_LOCAL, and move these conversion functions into the util module.