Add initial watchdog.py, which depends on a local Python install.
This version only notices viewer termination and distinguishes normal
termination from a crash. It also opens a watchdog.log in the viewer's logs
directory so we can tell. So far it logs successful termination as well as
detected crashes, but we'll eventually remove success-case logging.

Add an `LLProcessPtr mWatchdog` to `LLAppViewer`. Early in `LLAppViewer()`,
launch the watchdog process, passing it pipes for its stdin, stdout, stderr.
Of the three, the one that matters most is stdin: it's essential that the
watchdog receive a pipe from the viewer. That pipe will be closed by the OS on
viewer termination, one way or another.

Then in `~LLAppViewer()`, if the watchdog has been started and is still
running, write "OK" to it to reassure it that the viewer is terminating
normally.

Make viewer_manifest.py copy watchdog.py into the built viewer image.
Introduce Python's `pathlib.Path` into viewer_manifest.py, and defend
llmanifest.py against `Path` instances where it expects strings.
Clean up some redundancy in viewer_manifest.py.