Add a "run ID" to LLAppViewer, add "RunID" to static_debug_info.log.
This is to disambiguate reports sent multiple times for the same viewer run,
for crashes without a session ID.

Also add `ClientInfo/Channel` key to static_debug_info.log. This is identical
to `ClientInfo/Name`, but it might not be obvious to consumers of this data
packet that `Name` here means viewer channel. Nonetheless we leave `Name` in
place because it's likely that some existing analysis _does_ consult `Name`.

Also add `ClientInfo/Version` key. `ClientInfo` already contains
`MajorVersion`, `MinorVersion`, `PatchVersion` and `BuildVersion`, but why
make the consumer concatenate?