Fix watchdog "Date" and "Duration" calculations.
Turns out that on Windows, using st_birthtime or st_ctime gets you a
timestamp going back to the first viewer run on that filesystem. Use st_mtime,
which is current. But use st_mtime of SecondLife.start_marker instead of
SecondLife.log: the latter's st_mtime will be at the end of the run, not the
beginning. SecondLife.start_marker exists to record the beginning of a run.