Introduce workflow dispatch and mac functionality to qatest.yaml
A workflow dispatch has been added in an attempt to not only manually trigger this workflow but to also test this from a different branch without having to first merge to develop.

Also steps have been added to allow this workflow to run on mac runners when added. Mac runner info currently commented out.