LLTextureFetchWorker::doWork - Limit to half Decode Queue Length
Adding one was not enough to stop hitting the limit. Need to add the number of decode threads but no easy way to get that number and different viewers have different limits.

So, instead of adding a lot of code to find the current number of threads being used for decode, I just halved the queue limit from the full 1024 down to 512.

Hopefully, in the distant future when CPU's have 1000 cores, someone will see this commit and judge me kindly.