Fix for Mesh missing after TP. Assume MeshRezEnabled
The check in llvovolume for gMeshRepo.meshRezEnabled() has the side-effect of testing the region asset caps. After TP this is frequently failing and any mesh items in rebuildGeom at that moment will be excluded. Another option is to remove the check in llvovolume.cpp but as that was deliberately added as part of SL-5161 it is not clear whether that would be acceptable. Monty Linden confirmed that no current server version will decline that asset cap. Note that all this does is change the expected response to optimistic. If the region did return false subsequent updates would be treated correctly.