p#480 Collada vs GLTF Mesh Import Naming Convention
Both collada and gltf have a node and a mesh. Collada uses node-name, gltf was using mesh-name.
GLTF format permits reusing single mesh for multiple nodes, but nodes are warrantied to not be reused.

Switch to using node-names for better dupplicate avoidance and to be more in line with collada.